implement auto-drive handlers for upload, download, and search functionalities in the auto-mcp-servers package
